requirements.txt
streamlit>=1.24.0
langchain
langchain_ollama
langchain_openai
selenium>=4.10.0
beautifulsoup4>=4.9.3
lxml
html5lib
python-dotenv>=0.19.0
pandas>=1.5.0
requests>=2.28.0
google-generativeai
str
ibapi